Refactor: Outsource Funding reference event handler to its own module
- Moved Funding reference logic into a dedicated file (fundingreference.js)
- Replaced var with const for modern scoping
- Added full JSDoc documentation for clarity and maintainability
